backend/requirements.txt
# Automation ROI Calculator - Backend Dependencies
# =================================================

# Web Framework
fastapi==0.109.0          # Modern, fast web framework for building APIs
uvicorn==0.27.0           # ASGI server to run FastAPI applications

# Data Validation
pydantic==2.5.0           # Data validation using Python type hints

# PDF Generation
reportlab==4.0.0        # Uncomment when implementing PDF reports

# Database (Week 2)
# aiosqlite==0.19.0       # Uncomment when implementing database storage

# Authentication
python-jose==3.3.0